+++ lang/en/local_bulkenrol.php
@@ -41,7 +41,7 @@ $string['error_group_add_member'] = 'There was a problem when adding the user wi
 $string['error_group_add_members'] = 'There was a problem when adding the users to the course group(s).';
 $string['error_invalid_email'] = 'Invalid e-mail address found in line {$a->row} (<em>{$a->email}</em>). This line will be ignored.';
 $string['error_more_than_one_record_for_email'] = 'More than one existing Moodle user account with e-mail address <em>{$a}</em> found.<br />This line will be ignored, none of the existing Moodle users will be enrolled.';
-$string['error_no_courseid'] = 'No course id set.';
+$string['error_no_courseid'] = 'No course ID set.';
 $string['error_no_email'] = 'No e-mail address found in line {$a->line} (<em>{$a->content}</em>). This line will be ignored.';
 $string['error_no_record_found_for_email'] = 'No existing Moodle user account with e-mail address <em>{$a}</em>.<br />This line will be ignored, there won\'t be a Moodle user account created on-the-fly.';
 $string['error_no_valid_email_in_list'] = 'No valid e-mail address was found in the given list.<br />Please <a href=\'{$a->url}\'>go back and check your input</a>.';